verify: use int64_t in number_of_substrings

+++ verify/number_of_substrings.test.cpp
@@ -11,7 +11,7 @@ int main() {
 	std::string S; std::cin >> S;
 	int N = int(S.size());
 	auto sa = SuffixArrayLCP::shift_and_construct(S);
-	long long ans = (long long)(N) * (N+1) / 2;
+	int64_t ans = int64_t(N) * int64_t(N+1) / 2;
 	for (int i = 0; i < N; i++) ans -= sa.lcp[i];
 	std::cout << ans << '\n';
